+++ packages/common/src/decorators/injectable-decorator.ts
@@ -1,10 +1,9 @@
-import { DependencyContainer } from '../core';
 import { INJECT_MK, INJECTION_TOKEN_MK } from '../constants';
-import { ClassConstructor, Token } from '../types';
+import { ClassConstructor, InjectionToken, Token } from '../types';
 import 'reflect-metadata';
 export function Injectable() {
   return function (target: ClassConstructor) {
-    const token = DependencyContainer.createToken(target.name);
+    const token = new InjectionToken<any>(target.name);
     Reflect.defineMetadata(INJECTION_TOKEN_MK, token, target);
   };
 }