+++ packages/common/src/decorators/module-decorator.ts
@@ -1,15 +1,12 @@
-import { ModuleMetadata, ClassConstructor } from '../types';
+import { ModuleMetadata, ClassConstructor, InjectionToken } from '../types';
 import 'reflect-metadata';
 import { INJECTION_TOKEN_MK, ModuleMetadataKey } from '../constants';
-import { DependencyContainer } from 'core';
 
 export function Module(metadata: ModuleMetadata) {
   return function (target: ClassConstructor) {
     const processedProviders = metadata.providers?.map((provider) => {
       if (typeof provider === 'function') {
-        const token =
-          Reflect.getMetadata(INJECTION_TOKEN_MK, provider) ||
-          DependencyContainer.createToken(provider.name);
+        const token = Reflect.getMetadata(INJECTION_TOKEN_MK, provider) || new InjectionToken(provider.name);
         return {
           provide: token,
           useClass: provider,