+++ packages/core/src/index.ts
@@ -1,3 +1,1 @@
 export { Bunny } from './core';
-export * from './exceptions';
-export * from './decorators';